requirements.txt
Django==3.1.3
djangorestframework==3.12.2
django-filter==2.4.0
djangorestframework-simplejwt==4.5.0
Markdown==3.3.3
requests==2.24.0
django-nose==1.4.7
coverage==5.3